Upgrade Bolt from 1.2.9 to 1.2.15 to fix the multiplayer no longer working
There's a disclaimer here explaining why the multiplayer stopped working to begin with: https://doc.photonengine.com/bolt/current/getting-started/overview

I contacted Photon and they sent me version 1.2.15 of the Bolt SDK. I have no idea what the differences are between 1.2.15 and 1.2.14 since there are no release notes for 1.2.15, but it at least has the Google STUN server fix applied, so multiplayer works again now. I've tested the project with both server-client as well as dedicated server with clients and everything seems to work as it should.

I had to make 3 changes to the project's code to work with the new version of the SDK:

1. Update a namespace due to casing changes.
2. Change a method call to it's suggested replacement due to deprecation.
3. And I had to add add some specific handling for the "Launcher" scene in two places, due to what I assume are changes in the SDK, but I've commented this extensively in the code.